requirements.txt
pytesseract
Pillow
sentence-transformers
pymongo
uuid
requests
